MSL: Track SPIRVCrossDecorationInterfaceMemberIndex through nested structs.
SPIRVCrossDecorationInterfaceMemberIndex references the index of a member
of a variable that is a struct type. The value is relative to the variable,
and for structs nested within that top-level struct, the index value needs
to take into consideration the members within those nested structs.

Pass var_mbr_idx to add_plain_member_variable_to_interface_block() and
add_composite_member_variable_to_interface_block(), start at zero for each
variable, and increment for each member or nested member within that variable.

Add unit test shaders-msl/tese/in-block-with-nested-struct.tese.