Add special case to ignore offset decorator.
The offset decorator is not specifically only for SSBOs and UBOs,
they can be for any struct.

Specifically, there are some cases where the struct is only one member
long and it uses the Offset 0, which can be ignored since it doesn't
has any effect in the final struct layout.

We ignore that case but only for PODs (the ones that contains
scalar data).