[wip] MSL: Fix dynamically indexed pull interpolants
This is a first attempt at fixing the test failures mentioned in #1796
by copying scalarized/flattened arrays back into arrays in the fragment
shader for each "unique" interpolate opration. This approach has as
number of drawbacks such as negative impacts to runtime performance and
shader size, as well as having limited coverage for dynamic offsets
(InterpolateAtOffset) and samples (InterpolateAtSample).