MSL: Apply input override on all flattened members in interface block
As in #2305, the current logic cannot properly apply input override on
composite types in interface block. This change fixes the common case
where the composite types are flattened into the interface block.
The typing override is decoupled from location calculation so that it can
be applied to all members. In the meantime, some common logics are
consolidated.