MSL: Fix generation of the `WorkgroupSize` built-in.
Use any existing `gl_WorkGroupSize` when adding an implicit
`WorkgroupSize`. Make a constant using the `LocalSize` or `LocalSizeId`
when those execution modes are set. Only when none is present do we add
the `WorkgroupSize` as an input. Of course, since the `GLCompute` model
requires one of the three of constant `WorkgroupSize`, `LocalSize`, or
`LocalSizeId` to be present, this shouldn't happen in practice, but I
thought I'd handle it properly anyway.