MSL: Also handle signedness mismatches in fragment outputs.
It turns out that the number of components is not the only place where
Vulkan allows differences but Metal insists on strict equivalence.
Vulkan _also_ allows the fragment shader and the attachment to differ in
signedness--for example, rendering to an `R8G8B8A8_UINT` image where the
fragment shader declares its output to be an `ivec4`.

To fix this, we must adjust the `[[color]]` output in the interface
block to match the render target. Use the shader output specification
mechanism to do this; also move the fragment component count here to
save some memory at run time.

Signed-off-by: Chip Davis <cdavis5x@gmail.com>